fix(Movix): address metadata and presence review feedback
- update the metadata description to describe Movix itself instead of the activity
- anchor the URL regexp with an explicit https check
- replace custom presence typings with the PreMiD-provided types
- split the presence logic into multiple files for better readability
