+++ websites/M/Movix/routes/types.ts
@@ -1,0 +1,12 @@
+export interface RoutePresenceContext {
+  pathname: string
+  pageTitle: string
+  pageImage: string
+  contentImage: string
+  showTimestamp: boolean
+  showButtons: boolean
+}
+
+export type RoutePresenceHandler = (
+  context: RoutePresenceContext,
+) => Promise<PresenceData | null> | PresenceData | null
